requirements.txt
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
pyodbc==5.1.0
sqlite3==3.45.1.3 